+++ sema4ai/vscode-client/src/viewsRobots.ts
@@ -6,7 +6,6 @@ import * as roboCommands from "./robocorpCommands";
 import { basename, RobotEntry, RobotEntryType } from "./viewsCommon";
 import { getSelectedRobot } from "./viewsSelection";
 import { isActionPackage, isAgentPackage } from "./common";
-import { SEMA4AI_AGENT_PACKAGE_PUBLISH_TO_SEMA4_AI_STUDIO_APP } from "./robocorpCommands";
 
 let _globalSentMetric: boolean = false;
 
@@ -263,6 +262,15 @@ export class RobotsTreeDataProvider implements vscode.TreeDataProvider<RobotEntr
                         "iconPath": "go-to-file",
                         "type": RobotEntryType.OpenAgentSpecYaml,
                         "parent": element,
+                    },
+                    {
+                        "label": "Export Agent Package (zip)",
+                        "uri": element.uri,
+                        "robot": element.robot,
+                        "iconPath": "desktop-download",
+                        "type": RobotEntryType.ExportAgentPackage,
+                        "parent": element,
+                        "tooltip": "Exports the Agent Package as a ZIP archive",
                     },
                     {
                         "label": "Publish to Sema4.ai Studio",
@@ -630,6 +638,12 @@ export class RobotsTreeDataProvider implements vscode.TreeDataProvider<RobotEntr
             treeItem.command = {
                 "title": "Publish Agent Package to Sema4.ai Studio",
                 "command": roboCommands.SEMA4AI_AGENT_PACKAGE_PUBLISH_TO_SEMA4_AI_STUDIO_APP,
+            };
+            treeItem.collapsibleState = vscode.TreeItemCollapsibleState.None;
+        } else if (element.type === RobotEntryType.ExportAgentPackage) {
+            treeItem.command = {
+                "title": "Export Agent Package (zip)",
+                "command": roboCommands.SEMA4AI_PACK_AGENT_PACKAGE,
             };
             treeItem.collapsibleState = vscode.TreeItemCollapsibleState.None;
         } else if (element.type === RobotEntryType.PackageBuildToWorkspace) {